+++ tests/test_PIHandlerREST_connect.py
@@ -1,7 +1,8 @@
 import pytest
 import os
-import pandas as pd
-from tagreader.utils import datestr_to_datetime, ReaderType
+from tagreader.utils import (
+    ReaderType
+)
 from tagreader.web_handlers import (
     list_pi_servers,
     PIHandlerWeb,
@@ -17,7 +18,7 @@ if is_GITHUBACTION:
 
 
 BASE_URL = "https://pivision.equinor.com/piwebapi"
-ASSET = "PINO"
+ASSET = "PIMAM"
 TAGS = {
     "Float32": "BA:CONC.1",
     "Digital": "BA:ACTIVE.1",
@@ -55,32 +56,9 @@ def test_list_all_pi_servers():
         assert 3 <= len(r)
 
 
-def test_escape_chars():
-    assert (
-        PIHandlerWeb.escape('+-&|(){}[]^"~*:\\') == r"\+\-\&\|\(\)\{\}\[\]\^\"\~*\:\\"
-    )
-
-
 def test_verify_connection(PIHandler):
     assert PIHandler.verify_connection("PIMAM") is True
     assert PIHandler.verify_connection("somerandomstuffhere") is False
-
-
-def test_generate_search_query():
-    assert PIHandlerWeb.generate_search_query("SINUSOID") == {"q": "name:SINUSOID"}
-    assert PIHandlerWeb.generate_search_query(r"BA:*.1", server="PINO") == {
-        "q": r"name:BA\:*.1",
-        "scope": "pi:PINO",
-    }
-    assert PIHandlerWeb.generate_search_query(tag="BA:*.1") == {
-        "q": r"name:BA\:*.1",
-    }
-    assert PIHandlerWeb.generate_search_query(desc="Concentration Reactor 1") == {
-        "q": r"description:Concentration\ Reactor\ 1",
-    }
-    assert PIHandlerWeb.generate_search_query(
-        tag="BA:*.1", desc="Concentration Reactor 1"
-    ) == {"q": r"name:BA\:*.1 AND description:Concentration\ Reactor\ 1"}
 
 
 def test_search_tag(Client):
@@ -107,74 +85,6 @@ def test_tag_to_webid(PIHandler):
         res = PIHandler.tag_to_webid("SINUSOID*")
     with pytest.raises(AssertionError):
         res = PIHandler.tag_to_webid("somerandomgarbage")
-
-
-@pytest.mark.parametrize(
-    "read_type",
-    [
-        # pytest.param("RAW", marks=pytest.mark.skip(reason="Not implemented")),
-        # pytest.param(
-        #     "SHAPEPRESERVING", marks=pytest.mark.skip(reason="Not implemented")
-        # ),
-        "INT",
-        "MIN",
-        "MAX",
-        "RNG",
-        "AVG",
-        "STD",
-        "VAR",
-        # pytest.param("COUNT", marks=pytest.mark.skip(reason="Not implemented")),
-        # pytest.param("GOOD", marks=pytest.mark.skip(reason="Not implemented")),
-        # pytest.param("BAD", marks=pytest.mark.skip(reason="Not implemented")),
-        # pytest.param("TOTAL", marks=pytest.mark.skip(reason="Not implemented")),
-        # pytest.param("SUM", marks=pytest.mark.skip(reason="Not implemented")),
-        # pytest.param("SNAPSHOT", marks=pytest.mark.skip(reason="Not implemented")),
-    ],
-)
-def test_generate_read_query(PIHandler, read_type):  # TODO: Move away from test*connect
-    starttime = datestr_to_datetime(START_TIME)
-    stoptime = datestr_to_datetime(STOP_TIME)
-    ts = pd.Timedelta(SAMPLE_TIME, unit="s")
-
-    (url, params) = PIHandler.generate_read_query(
-        PIHandler.tag_to_webid("alreadyknowntag"),
-        starttime,
-        stoptime,
-        ts,
-        getattr(ReaderType, read_type),
-    )
-    assert params["startTime"] == "01-Apr-20 11:05:00"
-    assert params["endTime"] == "01-Apr-20 12:05:00"
-
-    if read_type == "INT":
-        assert url == f"streams/{PIHandler.webidcache['alreadyknowntag']}/interpolated"
-        assert params["interval"] == f"{SAMPLE_TIME}s"
-    elif read_type in ["AVG", "MIN", "MAX", "RNG", "STD", "VAR"]:
-        assert url == f"streams/{PIHandler.webidcache['alreadyknowntag']}/summary"
-        assert {
-            "AVG": "Average",
-            "MIN": "Minimum",
-            "MAX": "Maximum",
-            "RNG": "Range",
-            "STD": "StdDev",
-            "VAR": "StdDev",
-        }.get(read_type) == params["summaryType"]
-        assert params["summaryDuration"] == f"{SAMPLE_TIME}s"
-
-
-def test_is_summary(PIHandler):
-    assert PIHandler._is_summary(ReaderType.AVG)
-    assert PIHandler._is_summary(ReaderType.MIN)
-    assert PIHandler._is_summary(ReaderType.MAX)
-    assert PIHandler._is_summary(ReaderType.RNG)
-    assert PIHandler._is_summary(ReaderType.STD)
-    assert PIHandler._is_summary(ReaderType.VAR)
-    assert not PIHandler._is_summary(ReaderType.RAW)
-    assert not PIHandler._is_summary(ReaderType.SHAPEPRESERVING)
-    assert not PIHandler._is_summary(ReaderType.INT)
-    assert not PIHandler._is_summary(ReaderType.GOOD)
-    assert not PIHandler._is_summary(ReaderType.BAD)
-    assert not PIHandler._is_summary(ReaderType.SNAPSHOT)
 
 
 @pytest.mark.parametrize(
